# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Jaseo Job Tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pm/payouts:
    post:
      operationId: requestPmPayout
      tags: [pm-payouts]
      summary: Request a payout withdrawal (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PmPayoutInput"
      responses:
        "201":
          description: Payout request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmPayout"
        "400":
          description: Insufficient balance or below minimum
        "403":
          description: Forbidden
  /pm/payouts/mine:
    get:
      operationId: listMyPayouts
      tags: [pm-payouts]
      summary: List seller's payout requests
      responses:
        "200":
          description: Seller's payout requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PmPayout"
        "403":
          description: Forbidden
  /pm/payouts/admin:
    get:
      operationId: listAdminPayouts
      tags: [pm-admin]
      summary: List all payout requests (admin only)
      responses:
        "200":
          description: All payout requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PmPayout"
        "403":
          description: Forbidden
  /pm/payouts/{id}:
    patch:
      operationId: updatePmPayout
      tags: [pm-admin]
      summary: Approve or reject a payout (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PmPayoutUpdate"
      responses:
        "200":
          description: Updated payout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmPayout"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    PmPayout:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: ["string", "null"]
        sellerBankName:
          type: ["string", "null"]
        sellerBankAccount:
          type: ["string", "null"]
        sellerAccountOwner:
          type: ["string", "null"]
        amount:
          type: integer
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        transferProofKey:
          type: ["string", "null"]
        requestedAt:
          type: string
        processedAt:
          type: ["string", "null"]
      required:
        - id
        - sellerId
        - amount
        - status
        - requestedAt
    PmPayoutInput:
      type: object
      properties:
        amount:
          type: integer
      required:
        - amount
    PmPayoutUpdate:
      type: object
      properties:
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        transferProofKey:
          type: ["string", "null"]
      required:
        - status
...